R3000A: Fix Wsign-compare warning.

+++ pcsx2/R3000A.cpp
@@ -222,7 +222,7 @@ __ri void iopEventTest()
 	{
 		// start the next branch at the next counter event by default
 		// the interrupt code below will assign nearer branches if needed.
-		if (psxNextCounter < (psxRegs.iopNextEventCycle - psxNextsCounter))
+		if (psxNextCounter < static_cast<s32>(psxRegs.iopNextEventCycle - psxNextsCounter))
 			psxRegs.iopNextEventCycle = psxNextsCounter + psxNextCounter;
 	}
